+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project aims to follow [Semantic Versioning](https://semver.org/).
 
 - Added `messages.fetch_message_summary` for `GET /api/v2/domains/{domain}/messages/{messageId}/summary`.
 - Added `messages.fetch_message_text` for `GET /api/v2/domains/{domain}/messages/{messageId}/text`.
+- Added `messages.fetch_message_textplain` for `GET /api/v2/domains/{domain}/messages/{messageId}/textplain`.
 - Removed `wait` query parameter from `messages.fetch_sms_message` because the endpoint does not support it.
 - Removed `wait` query parameter from `messages.fetch_inbox`.
 - Fixed `webhooks.private_webhook` to include the webhook payload in the request body.